rename shared CSS module
